handle partial dedent
Here are 3 lines representing a "partial dedent":

1
  2
 3

2 is indented relative to 1, and 3 is dedented relative to 2, but not
enough to align with 1. In this case, we ignore the difference in
indentation between 2 and 3 and consider both to be indented by one
level relative to line 1.

This appears to be the meaning of this indentation on IOS 15. We can add
this behavior since partial dedent's behavior was essentially
unspecified before.